x86/mm: Add set_memory_r and set_memory_nr
Add set_memory_r() and set_memory_nr() functions for marking memory as not
readable. Tweak set_memory_ro() and set_memory_rw() to set the readable bit.

This is to enable execute-only memory, and so could be named set_memory_xo().
But there is already a set_memory_ro() and so calling set_memory_ro() and
set_memory_xo() together is confusing. Furthermore, set_memory_ro() doesn’t
actually do what it says (make memory _only_ have read permissions) because it
doesn’t touch NX. It should more properly be named set_memory_nw(). So to not
continue this trend, call setting the XO bit in the page tables as
set_memory_nr(). Also add a complimentary set_memory_r(), which will unset the
readable bit, but not touch the writable bit. This is combination is not
supported by KVM XO today, but it is a better API long term as other
architectures may support it or could emulate it.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>