modules: Add module support for XO
Add support for setting the text module section as XO and checking for a
supports_xo tag in the module. If the module is not tagged with “noxo” then
the text section will be set not readable when the entire module is set
read-only (module_enable_ro). The text will then be set readable again when the
module is set rw again.

Also, add a new function, set_all_modules_text_nr(), to be used by probing
later when text needs to be read to decode the probes instruction, but it’s
being patched via text_poke() so the it doesn’t need to set the module
mapping text writable.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>