x86/cpufeature: Add detection of KVM Capabilities
Add a new CPUID leaf to hold the contents of CPUID 0x40000030 EAX to detect KVM
defined generic VMM features.

The leaf was proposed to allow KVM to communicate features that are defined by
KVM, but available for any VMM to implement.

Add cpu_feature_enabled() support for features in this leaf (KVM XO), and a
pgtable_kvmxo_enabled() helper similar to pgtable_l5_enabled() so that
pgtable_kvmxo_enabled() can be used in early code that includes
arch/x86/include/asm/sparsemem.h.

Lastly, in head64.c detect and this feature and perform necessary adjustments
to physical_mask.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>