x86/mm: Add fault handler logic for XO violations
Add fault handler logic for XO violations that can handle setting the text back
to readable if config is set that way.

There are two ways of handling an XO fault, configured by
CONFIG_KVM_XO_KERNEL_ENFORCE. If CONFIG_KVM_XO_KERNEL_ENFORCE is set, the fault
will be an oops, like other kernel memory permission violations. If it is not
set, the fault handler will mark the faulting page as readable and resume
execution.

These faults may happen in an interrupt, and so no locks are taken when walking
the page tables. In order for this not to cause races.

1. No other permission changes can be made to not-readable text, except for
R->NR and NR->R if there is a chance it might be read.
2. None of the NR pages can be broken into smaller pages while there is a
chance they might be read.

All kernel memory needs to follow these rules whenever
CONFIG_KVM_XO_KERNEL_ENFORCE is not defined.

KVM XO does not support making writable but not readable memory. If such memory
is encountered the fault handler will not try to fix it. Instead it will cause
an oops.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>