x86/mm/cpa: Warn about invalid KVM XO permissions
When using the KVM virtualized XO bit, writeable but not readable
persissions are not possible. When the XO/NR bit is set, the underlying
EPT permissions will be XO, and so even if the writable bit is set in
the guest page tables, the page will not be writable. So warn if this
combination is ever encountered.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>