ftrace/x86: Use text_poke() for xo fixer case
This patch is based on an earlier patch to change ftrace to use the
text_poke() infrastructure. The point of that patch was to avoid leaving
RW mappings to text available for all CPUs while the patching was
happened. However, it was found that for some operations this was slower:
"Running:

  # time { echo function > current_tracer; }

takes 0.24s without this patch and 0.7s with."

And so it was not merged.

Separate from that justification, ftrace's patching races with the XO fault
fixer. If ftrace is patching the text at the same time the fault fixer
is fixing an NR page to be readable, one may overwrite the others page
table permisssion changes. To avoid this we can use text_poke() when the
XO fault fixer is active. By using a completely different RW mapping we
can avoid the fault fixer racing.

This patch is based off Nadav Amit's patch and so crediting him co-developed,
however he has not reviewed it.

Co-developed-by: Nadav Amit <namit@vmware.com>
Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>