KVM: Stop overloading USER to mean READ
Refactor the KVM MMU code to avoid overloading the ACC_USER_MASK
variable to mean readable in the EPT case. In upcoming patches support
will be added for guest MMUs to have memory that is both not readable
and user at the same time, so more ACC flags are needed. Add an
ACC_READ_MASK for this case, and as long as this now exists, make it
mean readable in the EPT case instead of ACC_USER_MASK.

To further disentangle this, add a new synthetic PFERR_ code to mean
readable access. Since EPT walks can involve simultaneous read and write
access in the case of dirty bits, it's not possible to distinguish
between a (write) vs a (write and read) by overloading present and
!write to signify a read fault in this case.

It could be possible to contintue to have PFERR_USER_MASK map to
ACC_READABLE in the EPT case but not in normal x86 page, but for clarity
use PFERR_READ_MASK for this case.

In picking a sythetic error code bit that is not yet defined, there
exists the small possibility that this bit will be used someday in real
and this bit overloading would need to be reworked. It is very unlikley
that such a high bit (31) will be used anytime soon as the SGX bit was
picked to be 15. So there is a lot of room to grow and time for KVM to
change if the bits start being used above the SGX bit. So in practice
this is very unlikley to cause a problem.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>