x86/cpufeature: Add detection of KVM PV XO
Detect KVM based execute-only (XO) feature and add a page type for it.
Do this by checking for the KVM XO CPUID bit and if present, reading the
XO position bit with the MSR. If it's present, then write the KVM XO MSR
to enable the feature. Define _page_xo and PAGE_EXECONLY. These only
have bits set if pgtable_pv_xo_enabled() is true.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>