KVM: x86/mmu: Add support for execute-only stolen bit
Add support for execute-only (XO) permissioned aliases controlled by a
new member of kvm_arch, xo_gpa_mask. This mask has only the XO bit set
and can be used to quickly determine if a GPA's stolen bits have the XO
bit set. Also add a gpa_stolen_mask to keep track of which stolen bits
have been enabled by the guest. Plug it into the existing helper
kvm_gpa_stolen_mask().

Also, use an ACC_EXEC permission if the XO stolen bit is set when
mapping a GPA.

Add a bit kvm_mmu_page_role to trigger rebuilding of the page tables.
The non-XO half of the page tables will have identical values as before
XO was enabled, and could in the future be shared to avoid re-building
when XO is turned on or off. Also add a member of kvm_vcpu_arch to hold
a state to allow switching the XO alias to RWX on a per-vcpu basis. This
can allow guests to easily toggle readability of XO mapped pages for a
single vcpu, which is useful in the case of PTEs shared between multiple
vcpus.

In addition, the lower levels of the page tables could be shared after
XO is enabled as well, with the XO permission only being set on an upper
level paging structure. These are both left for future optimizations.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>