KVM: Add logic for enabling/disabling KVM XO
Add the MSR for KVM to expose the execute-only (XO) bit and to enable
guests to turn on the feature. Opportunistically, align the values of
the other KVM MSRs.

Also, add logic for disabling XO permission enforcement on a per-vcpu
basis.

Add a new KVM request so that this can propagate to all vcpus by
resetting their mmu contexts.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>