KVM: Add KVM_CAP_EXEC_ONLY
Add KVM_CAP_EXEC_ONLY for host userspace to enable the execute-only (XO)
capability. This allows userspace to configure which PTE bit in the
guest is the XO permission bit.

In order for this feature to be supported, there are some restrictions
that need to be enforced:
 - The XO permission bit needs to be less than the host MAXPHYADDR, so
   that the alias can be mapped in the TDP.
 - The host needs to support XO TDP mappings.
 - The XO permission bit needs to be more than 36 to exclude paging
   modes except long and PAE.

Generic x86 KVM code needs to know if the hardware supports execute-only
memory for KVM XO. Add a function on kvm_x86_ops to expose this
capability.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>