x86/cpufeatures: Add execute-only feature bit
Some x86 CPUs have the ability to create execute only memory by
requesting only PROT_EXEC (without PROT_READ), however userspace has no
way of knowing if this memory will be execute-only (XO), or executable
and readable. With KVM XO memory, this kind of memory can be created by
VMs, so report this bit if that capability is available.

With protection keys for userspace, this kind of memory *can* be created
with general mmap calls, but not if the available protection keys have
been exhausted. So in the protection key's case, this kind of memory
won't be guaranteed to be created, and since execute-only memory is
often used for security usages, be conservative and don't select this
bit based on protection keys for userspace support.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>