mmap: Add support for PV XO
The paravirtualized execute only feature enables the ability to create
execute-only (XO) virtual memory. Enable the use of this feature for
userspace when PROT_EXEC and !PROT_READ, as the behavior in the case of
protection keys for userspace and previosly some arm64 platforms.

In the case of the ability to create execute-only memory with protection
keys AND the ability to create native execute-only memory, continue to
use the pkeys XO method of creating execute-only memory because of
potential slowdown when using PV XO method for some usages.

Set the values of the __P100 and __S100 in protection_map during boot
instead of statically because the actual PV XO bit in the PTE is
determinted at boot time and so can't be known at compile time.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>