x86/modules: Use real perm_allocations
x86 relocations will work as long as the relocation site is within 2GB.
As long as the allocations in somwhere in the module area, relocations can
be applied. So relax the restriction of having the module regions for a
module core or init area be virtually contiguous.

Also, apply relocations at the writable address of the perm_allocation to
support a future implementation that has the writable address in a
different allocation.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>