vmalloc: Add perm_alloc x86 implementation
Create a custom perm_alloc() implementation for x86. Allocate a new buffer
to be returned by perm_writable_addr(), and text_poke() it into the final
allocation in perm_writable_finish(). Map the allocation at it's final
permissions so that it does not need to be shot down as it transitions
from writable to read-only.

Since x86 direct map permissions are set via set_memory_() calls to alias
mappings that can be scattered across the direct map, instead draw from
blocks of 2MB unmapped pages. Register a shrinker callback to allow for
this pool of pages to be reclaimed. This should reduce breaking large
pages on the direct map over the existing set_memory_() based solutions.
Since the pages are left as 4k NP pages, we can remap them in the
reclaimer callback without any flush or split or flush.

By grabbing whole 2MB pages this may reduce the availability of high page
order blocks in some cases. However, it could just as easily improve the
availability of high order pages for cases where the vmalloc allocations
could be long lived, such as modules, by grouping these long lives
allocations together. By caching pages, this also reduces pages available
to other parts of the kernel for use in other caches.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>