KVM: x86/mmu: Support stolen bits being reserved in PTEs
The execute-only (XO) bit will used to create XO virtual memory by
setting the bit in the terminating guest page table entry, however the
guest could also set it in an entry that points to an intermediate PTE.
Unlike other permission bits, only the PTE that actually points to the
physical page being mapped will apply. Intermediate PTE's that point to
XO gpa aliases will cause EPT violations, due to the lack of R/W
permissions. So instead we will define the XO bit as reserved for PTE's
that point to other PTE's. Support this by allowing stolen bits to be
defined as reserved. This patch is preparatory, and will not result in
any new behavior until the XO bit is actually added.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>