mm: Add defines for protection_map[] indices
The protection_map[] is indexed via regular integers in places where it is
set at runtime. This is slightly hard to read as it's not easy to see
which numbers correspond to which permission combinations. In addition,
future patches involve changing protecion_map[] for individual isolated
permission combinations, making the integer indexes even more obscure.

Switch protection_map[] to use defines with names relevant to the
permission combinations for the index.

Suggested-by: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>