x86: Calculate MTRR masks from x86_phys_bits
Variable range MTRRs are currently determined with a physical address bit
size determined from logic that largely duplicates the logic used to
determine x86_phys_bits. For features that repurpose gpa bits for physical
memory aliases this duplication would only grow. So use x86_phys_bits
instead of recalculating.

Note, in order for this physical bit size calculation to be shared, any
x86_phys_bits adjustments need before the variable range MTRR masks are
configured. MKTME currently adjusts x86_phys_bits after MTRRs are set, but
there is no MKTME logic around MTRRs anyway so the behavior will not
change. If MKTME is ever implemented in the kernel, its x86_phys_bits
adjustment would need to be moved in order for this organization to work,
as currently it happens before the variable range MTRRs are calculated.

For SME/SEV, the x86_phys_bit is adjusted before the MTRR is set, so
now MTRRs that are set on a pfn that does not include the c-bit will
apply to both c-bit and non-c-bit gpa aliases.

TODO: Verify the correctness of the SME/SEV behavior change.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>